Minor tuning
