+++ lib/instance/instance_list.dart
@@ -1,0 +1,201 @@
+import 'package:dynamic_background/dynamic_background.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:provider/single_child_widget.dart';
+import 'package:smooth_list_view/smooth_list_view.dart';
+
+import '../settings/settings_model.dart';
+import '../setup/theme_setup.dart';
+import 'instance_manager.dart';
+import 'instance_model.dart';
+
+/// List of instances. This list shows all instances and allows the user to add
+/// and remove instances.
+class InstanceList extends StatelessWidget {
+  /// Constructor for the instance list.
+  const InstanceList({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Material(
+      color: Theme.of(context).colorScheme.surfaceContainer,
+      child: Consumer<InstanceModel>(
+        builder: (context, instances, child) {
+          return LayoutBuilder(
+            builder: (context, constraints) {
+              // TODO: Add break between tiles
+
+              return SmoothListView.builder(
+                key: const PageStorageKey('InstanceListSmoothListView'),
+                duration: const Duration(milliseconds: 500),
+                itemCount: instances.length + 1,
+                itemBuilder: (context, index) {
+                  if (index < instances.length) {
+                    return InstanceTile(
+                      index,
+                      instances[index],
+                      constraints.maxHeight,
+                    );
+                  }
+
+                  return Consumer<SettingsModel>(
+                    builder: (context, settings, child) {
+                      return ListTile(
+                        minTileHeight: 50,
+                        contentPadding:
+                            const EdgeInsets.symmetric(horizontal: 10),
+                        tileColor: settings.themeMode == ThemeMode.dark
+                            ? Theme.of(context).colorScheme.surfaceContainerHigh
+                            : Theme.of(context)
+                                .colorScheme
+                                .surfaceContainerHighest,
+                        leading: const Icon(
+                          Icons.add,
+                          size: 24,
+                        ),
+                        title: Text(
+                          'Add Instance',
+                          style: Theme.of(context).textTheme.bodyMedium,
+                        ),
+                        onTap: () async => instances.add(),
+                      );
+                    },
+                  );
+                },
+              );
+            },
+          );
+        },
+      ),
+    );
+  }
+}
+
+/// Tile for an instance. This tile shows the instance name and path. The tile
+/// changes color based on the instance's status.
+class InstanceTile extends StatelessWidget {
+  /// Constructor for the instance tile.
+  const InstanceTile(
+    int index,
+    InstanceController instance,
+    double maxHeight, {
+    super.key,
+  })  : _index = index,
+        _instance = instance,
+        _maxHeight = maxHeight;
+
+  final int _index;
+  final InstanceController _instance;
+  final double _maxHeight;
+
+  @override
+  Widget build(BuildContext context) {
+    final InstanceTheme instanceTheme =
+        Theme.of(context).extension<InstanceTheme>()!;
+
+    return Consumer<InstanceModel>(
+      builder: (context, instances, child) {
+        final isSelected = instances.selectedIndex == _index;
+
+        final Color selectedTileColor = Theme.of(context).colorScheme.secondary;
+
+        Color tileColor, hoverColor, selectedColor, warningColor;
+        if (_instance.isEnabled) {
+          tileColor = instanceTheme.enabled;
+          selectedColor = instanceTheme.enabled;
+          hoverColor = instanceTheme.enabledHover;
+          warningColor = instanceTheme.enabledWarning;
+        } else {
+          tileColor = instanceTheme.disabled;
+          selectedColor = instanceTheme.disabled;
+          hoverColor = instanceTheme.disabledHover;
+          warningColor = instanceTheme.disableWarning;
+        }
+
+        if (_instance.isNotValid) {
+          selectedColor = instanceTheme.warning;
+        }
+
+        return GestureDetector(
+          onSecondaryTapDown: (details) async => _showRemoveContextMenu(
+            context,
+            details.globalPosition,
+            removeInstance: () => instances.remove(_index),
+          ),
+          child: SingleChildBuilder(
+            builder: (context, child) {
+              if (_instance.isNotValid && !isSelected) {
+                return DynamicBg(
+                  height: 0,
+                  painterData: ScrollerPainterData(
+                    direction: ScrollDirection.left2Right,
+                    shape: ScrollerShape.stripesDiagonalForward,
+                    color: warningColor,
+                    backgroundColor: tileColor,
+                    fadeEdges: false,
+                  ),
+                  child: child,
+                );
+              } else {
+                return child!;
+              }
+            },
+            child: ListTile(
+              minTileHeight: 0,
+              contentPadding: const EdgeInsets.only(left: 10, right: 10),
+              tileColor: tileColor,
+              hoverColor: hoverColor,
+              selectedTileColor: selectedTileColor,
+              textColor: selectedTileColor,
+              selectedColor: selectedColor,
+              title: Text(
+                _instance.name,
+                maxLines: 1,
+                overflow: TextOverflow.ellipsis,
+              ),
+              subtitle: Text(
+                // Makes colon and space non-breaking
+                _instance.path
+                    .replaceFirst(':', ':\u2060')
+                    .replaceAll(' ', '\u202f'),
+                maxLines: _maxHeight > 350 ? 3 : (_maxHeight > 250 ? 2 : 1),
+                overflow: TextOverflow.ellipsis,
+              ),
+              onTap: () => instances.choose(_index),
+              selected: isSelected,
+            ),
+          ),
+        );
+      },
+    );
+  }
+
+  Future<void> _showRemoveContextMenu(
+    BuildContext context,
+    Offset position, {
+    required VoidCallback removeInstance,
+  }) async {
+    const remove = 'remove';
+
+    final String? result = await showMenu(
+      context: context,
+      position: RelativeRect.fromLTRB(
+        position.dx,
+        position.dy,
+        position.dx,
+        position.dy,
+      ),
+      menuPadding: EdgeInsets.zero,
+      popUpAnimationStyle: AnimationStyle.noAnimation,
+      items: [
+        const PopupMenuItem(
+          value: remove,
+          height: 40,
+          child: Text('Remove'),
+        ),
+      ],
+    );
+
+    if (result == remove) removeInstance();
+  }
+}
